pyproject.toml
[tool.poetry]
name = "alpaca-data-streams"
version = "0.1.0"
description = ""
authors = ["Andrea Mancuso <and.mancuso@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alpaca-py = "^0.30.1"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
simplejson = "^3.19.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
